backend/requirements.txt
fastapi
uvicorn
qdrant-client
google-genai
python-dotenv
openai





